HW 3.1 Исправлено все, кроме непонятных замечаний.

+++ hw03/src/ru/sberbank/homework/kudryavykh/TerminalServer.java
@@ -6,7 +6,8 @@ public class TerminalServer {
 
     public TerminalServer(long cardNumber) throws InvalidCardNumber {
         if (checkCardNumber(cardNumber)) {
-            account = new Account("User", 1000, (short) 1234, (short) 123, 1111222233334444L);
+            account = new Account("User", 1000, 1234, 123,
+                    1111222233334444L);
             pinValid = new PinValidator(account);
         } else {
             throw new InvalidCardNumber("Карта повреждена или не обслуживается");
